requirements.txt
pytesseract==0.3.10
Pillow==10.4.0
openai==1.52.0
python-dotenv==1.0.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
